requirements.txt
llama-index==0.10.18 
llama-index-llms-groq==0.1.3 
groq==0.4.2 
llama-index-embeddings-huggingface==0.2.0
flask
python-dotenv
IPython
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai